update Staggering Strikes for Midnight
